accessories_service/requirements.txt
flask
flask-cors
torch
torchvision
pandas
Pillow
scikit-learn
numpy